requirements.txt
beautifulsoup4==4.9.3
datasets==3.3.1
easyocr==1.7.1
google_search_results==2.4.2
huggingface_hub==0.31.2
langchain_anthropic==0.3.13
langchain_core==0.3.60
langchain_openai==0.3.17
llama_index==0.12.36
llama_parse==0.6.22
mammoth==1.9.0
markdownify==0.14.1
numpy==2.2.6
openai==1.79.0
opencv_python_headless==4.11.0.86
pandas==2.2.3
pathvalidate==3.2.3
pdf2image==1.17.0
pdfminer==20191125
pdfminer.six==20240706
Pillow==11.2.1
playwright==1.51.0
posthog==3.15.1
protobuf==4.25.3
puremagic==1.28
pydantic==2.11.4
pydub==0.25.1
PyPDF2==3.0.1
pytest==8.3.5
python-dotenv==1.1.0
python_pptx==1.0.2
ratelimit==2.2.1
Requests==2.32.3
rich==14.0.0
serpapi==0.1.5
setuptools==75.8.0
smolagents==1.15.0
streamlit==1.45.1
tabulate==0.9.0
torch==2.2.0
tqdm==4.67.1
transformers==4.49.0
translate==3.6.1
youtube_transcript_api==1.0.3
